+++ app_build_suite/__main__.py
@@ -62,6 +62,13 @@ def configure_global_options(config_parser: configargparse.ArgParser) -> None:
         type=BuildEngineType,
         help="Select the build engine used for building your chart.",
     )
+    config_parser.add_argument(
+        "--keep-going",
+        required=False,
+        default=False,
+        action="store_true",
+        help="Run all steps even if some fail, then report all errors at the end.",
+    )
     steps_group = config_parser.add_mutually_exclusive_group()
     steps_group.add_argument(
         "--steps",